fix: :bug: fix comment
- models: 일대다 관계의 매핑 타입 재설정
    - comment: 필드명 create_at을 created_at으로 수정
- review.schemas: type | None = None을 Optional[type]으로 수정
- comment.schemas: CommentCreate, Comment 스키마에서 review 필드 삭제
- comment.crud
    - create_comment: 필드명 create_at을 created_at으로 수정
    - get_comment_list 삭제제
- comment.router
    - comment_list 삭제
    - comment_update
        - 엔드포인트 수정
	- comment_id 파라미터 추가
    - comment_detail 추가
    - comment_delete
        - current_user 파라미터 추가
	- 작성자와 요청한 회원이 같지 않을 경우 403 오류 발생생

+++ api/domain/comment/router.py
@@ -1,5 +1,4 @@
 import sqlalchemy.orm as so
-from typing import List
 from fastapi import APIRouter, Depends, HTTPException
 from starlette import status
 from domain.account.router import load_current_account
@@ -12,17 +11,7 @@ from models import Account
 router = APIRouter(prefix="/api/comment")
 
 
-@router.get("/list", response_model=List[schemas.Comment])
-def comment_list(review_id: int, db: so.Session = Depends(get_db)):
-    review = r_crud.get_review(db, review_id)
-    if review is None:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND, detail="후기가 존재하지 않습니다."
-        )
-    return c_crud.get_comment_list(db, review)
-
-
-@router.post("/create", status_code=status.HTTP_204_NO_CONTENT)
+@router.post("/create/{review_id}", status_code=status.HTTP_204_NO_CONTENT)
 def comment_create(
     review_id: int,
     _comment_create: schemas.CommentCreate,
@@ -37,13 +26,12 @@ def comment_create(
     c_crud.create_comment(db, _comment_create, review, current_user)
 
 
-@router.put("/update/{comment_id}", status_code=status.HTTP_204_NO_CONTENT)
+@router.put("/update", status_code=status.HTTP_204_NO_CONTENT)
 def comment_update(
-    comment_id: int,
     _comment_create: schemas.CommentUpdate,
     db: so.Session = Depends(get_db),
 ):
-    comment = c_crud.get_comment(db, comment_id)
+    comment = c_crud.get_comment(db, _comment_create.comment_id)
     if comment is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="댓글이 존재하지 않습니다."
@@ -51,13 +39,22 @@ def comment_update(
     c_crud.update_comment(db, _comment_create, comment)
 
 
+@router.get("/detail/{comment_id}", response_model=schemas.Comment)
+def comment_detail(comment_id: int, db: so.Session = Depends(get_db)):
+    return c_crud.get_comment(db, comment_id)
+
+
 @router.delete("/delete", status_code=status.HTTP_204_NO_CONTENT)
 def comment_delete(
-    _comment_delete: schemas.CommentDelete, db: so.Session = Depends(get_db)
+    _comment_delete: schemas.CommentDelete,
+    db: so.Session = Depends(get_db),
+    current_user: Account = Depends(load_current_account),
 ):
     comment = c_crud.get_comment(db, _comment_delete.comment_id)
     if comment is None:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND, detail="댓글이 존재하지 않습니다."
         )
+    if current_user.id != comment.author.id:
+        raise HTTPException(status_code=status.HTTP_403, detail="삭제 권한이 없습니다.")
     c_crud.delete_comment(db, comment)
